Make WebsocketInterface a singleton.  Add shutdown hook to TesseractMain to clean up ws server, make it a groovy file
