+++ build.gradle
@@ -23,7 +23,7 @@ dependencies {
   compile group: 'org.slf4j', name: 'slf4j-api', version: '1.7.26'
 
   // https://mvnrepository.com/artifact/org.slf4j/slf4j-simple
-  compile group: 'org.slf4j', name: 'slf4j-simple', version: '1.7.26'
+  compile group: 'org.slf4j', name: 'slf4j-log4j12', version: '1.7.26'
 
   // Include all jar files in 'lib' automatically
   compile fileTree(dir: 'lib', include: ['**/*.jar'])
